feat(web): extract trust helpers into trust-lib
Move trust-reason and install-risk logic into trust-lib with expanded Vitest coverage.

Preserve the public @/lib/trust import surface.

Closes #4307
